refactor: Consolidate Apache HTTP clients into a single asynchronous implementation.
