+++ nodel-jyhost/build.gradle
@@ -73,9 +73,6 @@ dependencies {
     // for the Nodel HTTP client:
     implementation 'org.apache.httpcomponents.client5:httpclient5:5.2.1'
     implementation 'org.apache.httpcomponents.core5:httpcore5-reactive:5.2.1'
-    
-    // HTTP Components 4.x for backward compatibility
-    implementation 'org.apache.httpcomponents:httpclient:4.5.5'
 
     // convenience dependency: required *only* for 'recipe sync' node
     implementation 'org.eclipse.jgit:org.eclipse.jgit:3.7.1.+'
