+++ nodel-jyhost/src/main/java/org/nodel/http/impl/ApacheNodelHttpClient.java
@@ -1,177 +1,196 @@
 package org.nodel.http.impl;
 
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
 import java.net.InetAddress;
-import java.net.URI;
 import java.net.UnknownHostException;
 import java.security.SecureRandom;
-import java.security.cert.CertificateException;
 import java.security.cert.X509Certificate;
 import java.util.Map;
 import java.util.Map.Entry;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.TimeUnit;
 
 import javax.net.ssl.SSLContext;
 import javax.net.ssl.X509TrustManager;
 
-import org.apache.http.Header;
-import org.apache.http.HttpEntity;
-import org.apache.http.HttpEntityEnclosingRequest;
-import org.apache.http.HttpHost;
-import org.apache.http.HttpRequest;
-import org.apache.http.HttpResponse;
-import org.apache.http.ProtocolException;
-import org.apache.http.StatusLine;
-import org.apache.http.auth.AuthScope;
-import org.apache.http.auth.AuthenticationException;
-import org.apache.http.auth.Credentials;
-import org.apache.http.auth.NTCredentials;
-import org.apache.http.auth.UsernamePasswordCredentials;
-import org.apache.http.client.CredentialsProvider;
-import org.apache.http.client.RedirectStrategy;
-import org.apache.http.client.config.RequestConfig;
-import org.apache.http.client.methods.HttpDelete;
-import org.apache.http.client.methods.HttpEntityEnclosingRequestBase;
-import org.apache.http.client.methods.HttpGet;
-import org.apache.http.client.methods.HttpHead;
-import org.apache.http.client.methods.HttpOptions;
-import org.apache.http.client.methods.HttpPatch;
-import org.apache.http.client.methods.HttpPost;
-import org.apache.http.client.methods.HttpPut;
-import org.apache.http.client.methods.HttpRequestBase;
-import org.apache.http.client.methods.HttpTrace;
-import org.apache.http.client.methods.HttpUriRequest;
-import org.apache.http.config.Registry;
-import org.apache.http.config.RegistryBuilder;
-import org.apache.http.conn.socket.ConnectionSocketFactory;
-import org.apache.http.conn.socket.PlainConnectionSocketFactory;
-import org.apache.http.conn.ssl.NoopHostnameVerifier;
-import org.apache.http.conn.ssl.SSLConnectionSocketFactory;
-import org.apache.http.entity.StringEntity;
-import org.apache.http.impl.auth.BasicScheme;
-import org.apache.http.impl.client.BasicCredentialsProvider;
-import org.apache.http.impl.client.CloseableHttpClient;
-import org.apache.http.impl.client.HttpClientBuilder;
-import org.apache.http.impl.client.HttpClients;
-import org.apache.http.impl.conn.PoolingHttpClientConnectionManager;
-import org.apache.http.protocol.HttpContext;
+import org.apache.hc.client5.http.async.methods.SimpleHttpRequest;
+import org.apache.hc.client5.http.async.methods.SimpleHttpResponse;
+import org.apache.hc.client5.http.async.methods.SimpleRequestBuilder;
+import org.apache.hc.client5.http.auth.AuthScope;
+import org.apache.hc.client5.http.auth.Credentials;
+import org.apache.hc.client5.http.auth.NTCredentials;
+import org.apache.hc.client5.http.auth.UsernamePasswordCredentials;
+import org.apache.hc.client5.http.config.RequestConfig;
+import org.apache.hc.client5.http.impl.DefaultRedirectStrategy;
+import org.apache.hc.client5.http.impl.async.CloseableHttpAsyncClient;
+import org.apache.hc.client5.http.impl.async.HttpAsyncClients;
+import org.apache.hc.client5.http.impl.auth.BasicCredentialsProvider;
+import org.apache.hc.client5.http.impl.nio.PoolingAsyncClientConnectionManager;
+import org.apache.hc.client5.http.impl.nio.PoolingAsyncClientConnectionManagerBuilder;
+import org.apache.hc.client5.http.ssl.ClientTlsStrategyBuilder;
+import org.apache.hc.client5.http.ssl.NoopHostnameVerifier;
+import org.apache.hc.core5.concurrent.FutureCallback;
+import org.apache.hc.core5.http.ContentType;
+import org.apache.hc.core5.http.Header;
+import org.apache.hc.core5.http.HttpHost;
+import org.apache.hc.core5.http.HttpRequest;
+import org.apache.hc.core5.http.HttpResponse;
+import org.apache.hc.core5.http.protocol.HttpContext;
+import org.apache.hc.core5.io.CloseMode;
+import org.apache.hc.core5.reactor.IOReactorConfig;
+import org.apache.hc.core5.ssl.SSLContexts;
+import org.apache.hc.core5.util.Timeout;
 import org.nodel.Strings;
 import org.nodel.Version;
-import org.nodel.diagnostics.CountableInputStream;
-import org.nodel.diagnostics.SharableMeasurementProvider;
-import org.nodel.io.Stream;
-import org.nodel.io.UnexpectedIOException;
 import org.nodel.net.HTTPSimpleResponse;
 import org.nodel.net.NodelHTTPClient;
 
+/**
+ * Apache HTTP Client 5 implementation of NodelHTTPClient that supports asynchronous HTTP requests.
+ */
 public class ApacheNodelHttpClient extends NodelHTTPClient {
     
-    private Object _lock = new Object();
+    private final Object _lock = new Object();
     
-    /**
-     * The Apache Http Client
-     * (created lazily)
-     */
-    private CloseableHttpClient _httpClient;
-    
-    /**
-     * Required with 'applySecurity'
-     */
-    private CredentialsProvider _credentialsProvider;
-
-    /**
-     * Mainly used for adjusting timeouts
-     */
+    private CloseableHttpAsyncClient _httpAsyncClient;
+    private ExecutorService _executor;
+    private BasicCredentialsProvider _credentialsProvider;
     private RequestConfig _requestConfig;
     
     /**
      * Maximum content allowed to avoid uncontrolled memory allocation
-     * (default 150 MB, anything more should be using a streaming technique instead)
      */
     private final static int MAX_ALLOWED = 150 * 1024 * 1024;
     
     /**
-     * This needs to be done lazily because proxy can only be set up once
-     * 
-     * (uses double-check singleton)
+     * Ignores all SSL verification
      */
-    private void lazyInit() {
-        if (_httpClient == null) {
-            synchronized (_lock) {
-                if (_httpClient != null)
-                    return;
-                
-                HttpClientBuilder builder = HttpClients.custom()
-                        .setUserAgent("Nodel/" + Version.shared().version)
-                        
-                        // need to reference this later
-                        .setDefaultCredentialsProvider(_credentialsProvider = new BasicCredentialsProvider())
-                        
-                        // unrestricted connections
-                        .setMaxConnTotal(1000)
-                        .setMaxConnPerRoute(1000)
-                       
-                        // default timeouts
-                        .setDefaultRequestConfig(_requestConfig = RequestConfig.custom()
-                                .setConnectTimeout(DEFAULT_CONNECTTIMEOUT)
-                                .setSocketTimeout(DEFAULT_READTIMEOUT)
-                                .build());
-                
-                // using a proxy?
-                if (!Strings.isBlank(_proxyAddress))
-                    builder.setProxy(prepareForProxyUse(_proxyAddress, _proxyUsername, _proxyPassword));
-                
-                // ignore all SSL verifications errors?
-                if (_ignoreSSL)
-                    prepareForNoSSL(builder);
-                
-                // ignore all redirect codes
-                if (_ignoreRedirects)
-                    builder.setRedirectStrategy(IGNORE_ALL_REDIRECTS);
-                
-                // build the client
-                _httpClient = builder.build();
+    private static final X509TrustManager IGNORE_SSL_TRUSTMANAGER = new X509TrustManager() {
+        public void checkClientTrusted(X509Certificate[] chain, String authType) {}
+        public void checkServerTrusted(X509Certificate[] chain, String authType) {}
+        public X509Certificate[] getAcceptedIssuers() { return new X509Certificate[0]; }
+    };
+    
+    /**
+     * Ignores all redirect directives (will be manually handled)
+     */
+    private static final DefaultRedirectStrategy IGNORE_ALL_REDIRECTS = new DefaultRedirectStrategy() {
+        @Override
+        public boolean isRedirected(HttpRequest request, HttpResponse response, HttpContext context) {
+            return false;
+        }
+    };
+    
+    /**
+     * Ensure the asynchronous client is initialized
+     */
+    private synchronized void initAsyncClient() {
+        if (_httpAsyncClient != null) return;
+        
+        // Base configuration
+        if (_credentialsProvider == null) {
+            _credentialsProvider = new BasicCredentialsProvider();
+            _requestConfig = RequestConfig.custom()
+                    .setConnectTimeout(Timeout.ofMilliseconds(DEFAULT_CONNECTTIMEOUT))
+                    .setResponseTimeout(Timeout.ofMilliseconds(DEFAULT_READTIMEOUT))
+                    .build();
+        }
+        
+        // Set up executor
+        if (_executor == null) {
+            _executor = Executors.newCachedThreadPool();
+        }
+        
+        // Configure IO reactor
+        IOReactorConfig ioReactorConfig = IOReactorConfig.custom()
+                .setSoTimeout(Timeout.ofMilliseconds(DEFAULT_READTIMEOUT))
+                .build();
+        
+        org.apache.hc.client5.http.impl.async.HttpAsyncClientBuilder builder = HttpAsyncClients.custom()
+                .setUserAgent("Nodel/" + Version.shared().version)
+                .setDefaultCredentialsProvider(_credentialsProvider)
+                .setDefaultRequestConfig(_requestConfig)
+                .setIOReactorConfig(ioReactorConfig);
+        
+        // Configure pooling and SSL
+        PoolingAsyncClientConnectionManagerBuilder poolBuilder = PoolingAsyncClientConnectionManagerBuilder.create()
+                .setMaxConnTotal(1000)
+                .setMaxConnPerRoute(1000);
+
+        if (_ignoreSSL) {
+            try {
+                SSLContext sslContext = SSLContexts.custom().loadTrustMaterial(null, (chain, authType) -> true).build();
+                poolBuilder.setTlsStrategy(ClientTlsStrategyBuilder.create()
+                        .setSslContext(sslContext)
+                        .setHostnameVerifier(NoopHostnameVerifier.INSTANCE)
+                        .build());
+            } catch (Exception e) {
+                throw new RuntimeException("Error initializing SSL context", e);
             }
         }
-    }
+        
+        builder.setConnectionManager(poolBuilder.build());
 
-    /**
-     * (convenience method) 
-     */
-    private HttpHost prepareForProxyUse(String proxyAddress, String proxyUsername, String proxyPassword) {
-        HttpHost proxy;
-        
-        String proxyHost = null;
-        int proxyPort = -1;
-        
-        try {
-            int lastIndexOfColon = proxyAddress.lastIndexOf(':');
-            proxyHost = proxyAddress.substring(0, lastIndexOfColon);
-            proxyPort = Integer.parseInt(proxyAddress.substring(lastIndexOfColon + 1));
-        } catch (Exception ignore) {
+        // Configure proxy if needed
+        if (!Strings.isBlank(_proxyAddress)) {
+            builder.setProxy(createProxyHost(_proxyAddress, _proxyUsername, _proxyPassword));
         }
         
-        if (Strings.isBlank(proxyHost) || proxyPort <= 0)
-            throw new IllegalArgumentException("Proxy address is not in form host:port");
+        // Configure redirects if needed
+        if (_ignoreRedirects) {
+            builder.setRedirectStrategy(IGNORE_ALL_REDIRECTS);
+        }
         
-        proxy = new HttpHost(proxyHost, proxyPort);
+        _httpAsyncClient = builder.build();
+        _httpAsyncClient.start();
+    }
+    
+    /**
+     * Create and configure a proxy host
+     */
+    private HttpHost createProxyHost(String proxyAddress, String proxyUsername, String proxyPassword) {
+        int lastIndexOfColon = proxyAddress.lastIndexOf(':');
+        if (lastIndexOfColon <= 0) {
+            throw new IllegalArgumentException("Proxy address must be in the form host:port");
+        }
         
-        // using proxy credentials?
+        String proxyHost = proxyAddress.substring(0, lastIndexOfColon);
+        int proxyPort;
+        
+        try {
+            proxyPort = Integer.parseInt(proxyAddress.substring(lastIndexOfColon + 1));
+        } catch (NumberFormatException e) {
+            throw new IllegalArgumentException("Invalid proxy port");
+        }
+        
+        if (Strings.isBlank(proxyHost) || proxyPort <= 0) {
+            throw new IllegalArgumentException("Proxy address must be in the form host:port");
+        }
+        
+        HttpHost proxy = new HttpHost(proxyHost, proxyPort);
+        
+        // Configure proxy credentials if provided
         if (!Strings.isBlank(proxyUsername) && proxyPassword != null) {
             String userPart = proxyUsername;
             String domainPart = null;
             int indexOfBackSlash = proxyUsername.indexOf('\\');
+            
             if (indexOfBackSlash > 0) {
                 domainPart = proxyUsername.substring(0, indexOfBackSlash);
                 userPart = proxyUsername.substring(Math.min(proxyUsername.length() - 1, indexOfBackSlash + 1));
             }
+            
             AuthScope authScope = new AuthScope(proxyHost, proxyPort);
+            
             if (domainPart == null) {
-                _credentialsProvider.setCredentials(authScope, new UsernamePasswordCredentials(proxyUsername, proxyPassword));
+                _credentialsProvider.setCredentials(authScope, 
+                        new UsernamePasswordCredentials(proxyUsername, proxyPassword.toCharArray()));
             } else {
-                // normally used with NTLM
-                _credentialsProvider.setCredentials(authScope, new NTCredentials(userPart, proxyPassword, getLocalHostName(), domainPart));
+                _credentialsProvider.setCredentials(authScope, 
+                        new NTCredentials(userPart, proxyPassword.toCharArray(), getLocalHostName(), domainPart));
             }
         }
         
@@ -179,28 +198,15 @@ public class ApacheNodelHttpClient extends NodelHTTPClient {
     }
     
     /**
-     * (convenience method)
+     * Returns the local host name (without exceptions)
      */
-    private void prepareForNoSSL(HttpClientBuilder builder) {
+    private static String getLocalHostName() {
         try {
-            SSLContext sslContext = SSLContext.getInstance("TLS");
-            sslContext.init(null, new X509TrustManager[] { IGNORE_SSL_TRUSTMANAGER }, new SecureRandom());
-            builder.setSSLContext(sslContext);
-            
-            SSLConnectionSocketFactory sslSocketFactory = new SSLConnectionSocketFactory(sslContext, new NoopHostnameVerifier());
-            builder.setSSLSocketFactory(sslSocketFactory);
-            
-            Registry<ConnectionSocketFactory> socketFactoryRegistry = RegistryBuilder.<ConnectionSocketFactory>create()
-                    .register("http", PlainConnectionSocketFactory.getSocketFactory())
-                    .register("https", sslSocketFactory)
-                    .build();
-            
-            PoolingHttpClientConnectionManager connMgr = new PoolingHttpClientConnectionManager(socketFactoryRegistry);
-            builder.setConnectionManager(connMgr);
-        } catch (Exception e) {
-            throw new RuntimeException(e);
+            return InetAddress.getLocalHost().getHostName();
+        } catch (UnknownHostException e) {
+            return "localhost";
         }
-    }    
+    }
     
     @Override
     public HTTPSimpleResponse makeRequest(String urlStr, String method, Map<String, String> query, 
@@ -208,330 +214,265 @@ public class ApacheNodelHttpClient extends NodelHTTPClient {
                          Map<String, String> headers, String contentType, 
                          String body, 
                          Integer connectTimeout, Integer readTimeout) {
-        lazyInit();
+        try {
+            // Delegate to async implementation for consistency
+            return makeRequestAsync(urlStr, method, query, username, password, 
+                    headers, contentType, body, connectTimeout, readTimeout).get();
+        } catch (InterruptedException e) {
+            Thread.currentThread().interrupt();
+            throw new RuntimeException("Request was interrupted", e);
+        } catch (ExecutionException e) {
+            Throwable cause = e.getCause();
+            if (cause instanceof RuntimeException) {
+                throw (RuntimeException) cause;
+            }
+            throw new RuntimeException("Error executing request", cause);
+        }
+    }
+    
+    @Override
+    public CompletableFuture<HTTPSimpleResponse> makeRequestAsync(String urlStr, String method, Map<String, String> query, 
+                         String username, String password, 
+                         Map<String, String> headers, String contentType, 
+                         String body, 
+                         Integer connectTimeout, Integer readTimeout) {
         
-        // record rate of new connections
+        // Initialize the async client
+        synchronized (_lock) {
+            initAsyncClient();
+        }
+        
+        // Record metrics
         s_attemptRate.incrementAndGet();
         
-        // construct the full URL (includes query string)
-        String fullURL;
-        
+        // Build full URL with query parameters
+        String fullURL = urlStr;
         String queryPart = urlEncodeQuery(query);
-        if (!Strings.isEmpty(queryPart))
+        if (!Strings.isEmpty(queryPart)) {
             fullURL = String.format("%s?%s", urlStr, queryPart);
-        else
-            fullURL = urlStr;
+        }
 
-        // (out of scope for clean up purposes)
-        InputStream inputStream = null;
-        CountableInputStream cis = null;
-        
-        boolean executed = false;
+        // Create the CompletableFuture to return
+        CompletableFuture<HTTPSimpleResponse> future = new CompletableFuture<>();
         
         try {
-            // GET, POST or other
-            HttpRequestBase request = selectHTTPbyMethod(method, body, fullURL);
+            // Build the request
+            SimpleHttpRequest request = createRequest(method, fullURL, body, contentType);
             
-            // if username is supplied, apply security
-            if (!Strings.isBlank(username))
-                applySecurity(request, username, !Strings.isEmpty(password) ? password : "");            
+            // Apply authentication if provided
+            if (!Strings.isBlank(username)) {
+                applyAuth(request, username, !Strings.isEmpty(password) ? password : "");
+            }
             
-            // set 'Content-Type' header
-            if (!Strings.isBlank(contentType))
-                request.setHeader("Content-Type", contentType);
-
-            // add (or override) any request headers
+            // Add headers
             if (headers != null) {
-                for (Entry<String, String> entry : headers.entrySet())
+                for (Entry<String, String> entry : headers.entrySet()) {
                     request.setHeader(entry.getKey(), entry.getValue());
+                }
             }
             
-            if (!Strings.isEmpty(body)) {
-                if (!(request instanceof HttpEntityEnclosingRequest))
-                    throw new IllegalArgumentException("The HTTP method does not accept a body - " + method);
-                
-                HttpEntityEnclosingRequest httpWithBody = (HttpEntityEnclosingRequest) request;
-                httpWithBody.setEntity(new StringEntity(body, "utf-8"));
-            }
+            // Apply per-request timeouts
+            applyTimeouts(request, connectTimeout, readTimeout);
             
-            // set any timeouts that apply
-            if (connectTimeout != null || readTimeout != null) {
-                int actualConnTimeout = connectTimeout != null ? connectTimeout : DEFAULT_CONNECTTIMEOUT;
-                int actualReadTimeout = readTimeout != null ? readTimeout : DEFAULT_READTIMEOUT;
-
-                request.setConfig(RequestConfig.copy(_requestConfig)
-                        .setConnectTimeout(actualConnTimeout)
-                        .setSocketTimeout(actualReadTimeout)
-                        .build());
-            }
-            
-            // perform the request 
-            
-            // (and count it)
-            executed = true;
+            // Track active connections
             s_activeConnections.incrementAndGet();
-            
-            HttpResponse httpResponse = _httpClient.execute(request);
-            
-            // count the post now
-            if (!Strings.isEmpty(body))
-                s_sendRate.addAndGet(body.length());
-            
-            // safely get the response (regardless of response code for now)
-            String content = null;
-            
-            // safely get the content encoding
-            HttpEntity entity = httpResponse.getEntity();
-            if (entity != null) {
-                Header contentEncodingHeader = entity.getContentEncoding();
-                String contentEncoding = null;
-                if (contentEncodingHeader != null)
-                    contentEncoding = contentEncodingHeader.getValue();
 
-                // only if no specific encoding specified, fallback to UTF-8 for json and xml
-                if (contentEncoding == null) {
-                    Header recvContentTypeHeader = entity.getContentType();
-                    if (recvContentTypeHeader != null) {
-                        String recvContentType = recvContentTypeHeader.getValue() != null ? recvContentTypeHeader.getValue().toLowerCase() : ""; // as safe lower-case
+            // Execute the request
+            _httpAsyncClient.execute(request, new FutureCallback<SimpleHttpResponse>() {
+                @Override
+                public void completed(SimpleHttpResponse response) {
+                    try {
+                        // Count metrics
+                        if (!Strings.isEmpty(body)) {
+                            s_sendRate.addAndGet(body.length());
+                        }
+                        
+                        byte[] contentBytes = response.getBodyBytes();
+                        if (contentBytes != null) {
+                            if (contentBytes.length > MAX_ALLOWED) {
+                                throw new IOException("Too big - HTTP response over " + MAX_ALLOWED + " bytes is not allowed");
+                            }
+                            s_receiveRate.addAndGet(contentBytes.length);
+                        }
+                        
+                        String content = null;
+                        if (contentBytes != null) {
+                            String encoding = "ISO-8859-1"; // Default
+                            ContentType contentType = response.getContentType();
+                            if (contentType != null) {
+                                if (contentType.getCharset() != null) {
+                                    encoding = contentType.getCharset().name();
+                                } else {
+                                    String mimeType = contentType.getMimeType().toLowerCase();
+                                    if (mimeType.contains("json") || mimeType.contains("xml")) {
+                                        encoding = "utf-8";
+                                    }
+                                }
+                            }
+                            content = new String(contentBytes, encoding);
+                        }
 
-                        if (recvContentType.contains("json") || recvContentType.contains("xml"))
-                            contentEncoding = "utf-8";
+                        // Create response object
+                        HTTPSimpleResponse result = new HTTPSimpleResponse();
+                        result.content = content;
+                        result.statusCode = response.getCode();
+                        result.reasonPhrase = response.getReasonPhrase();
+                        
+                        // Add response headers
+                        for (Header header : response.getHeaders()) {
+                            result.addHeader(header.getName(), header.getValue());
+                        }
+                        
+                        future.complete(result);
+                    } catch (Exception e) {
+                        failed(e);
+                    } finally {
+                        s_activeConnections.decrementAndGet();
                     }
                 }
-
-                inputStream = entity.getContent();
-
-                // inject a counter so bytes can be counted, not characters
-                SafeCounter byteCounter = new SafeCounter();
-                cis = new CountableInputStream(inputStream, SharableMeasurementProvider.Null.INSTANCE, byteCounter);
-
-                // try using the given encoding or a straight 8-bit widening (for convenience ISO-8859-1 does that trick)
-                String encodingToUse = Strings.isBlank(contentEncoding) ? "ISO-8859-1" : contentEncoding;
-
-                InputStreamReader isr = new InputStreamReader(cis, encodingToUse); // unknown encoding will raise an exception
-                content = Stream.readFully(isr);
-            }
-            
-            StatusLine statusLine = httpResponse.getStatusLine();
-            
-            HTTPSimpleResponse result = new HTTPSimpleResponse();
-            result.content = content;
-            result.statusCode = statusLine.getStatusCode();
-            result.reasonPhrase = statusLine.getReasonPhrase();
-            
-            for (Header header : httpResponse.getAllHeaders())
-                result.addHeader(header.getName(), header.getValue());
-            
-            return result;
-            
-        } catch (IOException exc) {
-            throw new UnexpectedIOException(exc);
-            
-        } finally {
-            Stream.safeClose(inputStream);
-            
-            // count how many bytes were processed (failure or not)
-            if (cis != null)
-                s_receiveRate.addAndGet(cis.getTotal());
-            
-            if (executed)
-                s_activeConnections.decrementAndGet();
-        }
-    }
-
-    /**
-     * (convenience function)
-     */
-    private static HttpRequestBase selectHTTPbyMethod(String method, String body, String url) {
-        // deal with in order of likelihood
-        if (Strings.isBlank(method) && Strings.isEmpty(body) || "GET".equals(method))
-            return new HttpGet(url);
-        
-        else if (Strings.isBlank(method) && !Strings.isEmpty(body) || "POST".equals(method))
-            return new HttpPost(url);
-        
-        else if ("DELETE".equals(method))
-            return new HttpDelete(url);
-        
-        else if ("HEAD".equals(method))
-            return new HttpHead(url);
-        
-        else if ("PUT".equals(method))
-            return new HttpPut(url);
-        
-        else if ("TRACE".equals(method))
-            return new HttpTrace(url);
-        
-        else if ("OPTIONS".equals(method))
-            return new HttpOptions(url);
-
-        else if ("PATCH".equals(method))
-            return new HttpPatch(url);
-
-        else if (!Strings.isBlank(method))
-            return nonstandardHTTPMethod(method, body, url);
-
-        else
-            throw new IllegalArgumentException("Unknown HTTP method - " + method);
-    }
-
-    /**
-     * (convenience function)
-     */
-    private static HttpRequestBase nonstandardHTTPMethod(String method, String body, String url) {
-        // nonstandard / uncommon method has been specified - related issue #344
-        HttpRequestBase request;
-
-        // select a method class that does or does not support content (enclosed entity)
-        if (Strings.isEmpty(body)) {
-            request = new HttpRequestBase() {
-
+                
                 @Override
-                public String getMethod() {
-                    return method;
+                public void failed(Exception ex) {
+                    try {
+                        future.completeExceptionally(new IOException(ex));
+                    } finally {
+                        s_activeConnections.decrementAndGet();
+                    }
                 }
-
-            };
-        } else {
-            request = new HttpEntityEnclosingRequestBase() {
-
+                
                 @Override
-                public String getMethod() {
-                    return method;
+                public void cancelled() {
+                    try {
+                        future.completeExceptionally(new RuntimeException("Request was cancelled"));
+                    } finally {
+                        s_activeConnections.decrementAndGet();
+                    }
                 }
-
-            };
+            });
+        } catch (Exception e) {
+            future.completeExceptionally(new IOException(e));
         }
-        request.setURI(URI.create(url));
-        return request;
+        
+        return future;
     }
-
+    
     /**
-     * Applies security for a given HTTP request.
-     * @throws AuthenticationException 
+     * Apply custom timeouts to a request
      */
-    private void applySecurity(HttpRequestBase httpRequest, String username, String password) {
-        Credentials creds;
+    private void applyTimeouts(SimpleHttpRequest request, Integer connectTimeout, Integer readTimeout) {
+        int actualConnTimeout = connectTimeout != null ? connectTimeout : DEFAULT_CONNECTTIMEOUT;
+        int actualReadTimeout = readTimeout != null ? readTimeout : DEFAULT_READTIMEOUT;
 
-        // in case of NTLM, check for '\' in username
-        String userPart = username;
-        String domainPart = null;
-        int indexOfBackSlash = username.indexOf('\\');
-        if (indexOfBackSlash > 0) {
-            // NTLM
-            domainPart = username.substring(0, indexOfBackSlash);
-            userPart = username.substring(Math.min(username.length() - 1, indexOfBackSlash + 1));
-
-            creds = new NTCredentials(userPart, password, getLocalHostName(), domainPart);
-
-        } else {
-            // BasicAuth
-            creds = new UsernamePasswordCredentials(username, password);
-
-            // pre-emptive
-            try {
-                httpRequest.setHeader(new BasicScheme().authenticate(creds, httpRequest, null));
-            } catch (AuthenticationException e) {
-                throw new RuntimeException(e);
+        RequestConfig customConfig = RequestConfig.copy(_requestConfig)
+                .setConnectTimeout(Timeout.ofMilliseconds(actualConnTimeout))
+                .setResponseTimeout(Timeout.ofMilliseconds(actualReadTimeout))
+                .build();
+        
+        request.setConfig(customConfig);
+    }
+    
+    /**
+     * Apply authentication to a request
+     */
+    private void applyAuth(SimpleHttpRequest request, String username, String password) {
+        // Add Basic Auth header directly to the request
+        request.setHeader("Authorization", "Basic " +
+                java.util.Base64.getEncoder().encodeToString((username + ":" + password).getBytes()));
+        
+        // Set up NTLM auth if domain is specified with backslash
+        try {
+            String userPart = username;
+            String domainPart = null;
+            int indexOfBackSlash = username.indexOf('\\');
+            
+            if (indexOfBackSlash > 0) {
+                domainPart = username.substring(0, indexOfBackSlash);
+                userPart = username.substring(Math.min(username.length() - 1, indexOfBackSlash + 1));
+                
+                Credentials creds = new NTCredentials(userPart, password.toCharArray(), getLocalHostName(), domainPart);
+                _credentialsProvider.setCredentials(
+                        new AuthScope(request.getAuthority().getHostName(), request.getAuthority().getPort()),
+                        creds);
+            } else {
+                Credentials creds = new UsernamePasswordCredentials(username, password.toCharArray());
+                _credentialsProvider.setCredentials(
+                        new AuthScope(request.getAuthority().getHostName(), request.getAuthority().getPort()),
+                        creds);
             }
+        } catch (Exception e) {
+            // If credential setting fails, continue anyway as the header is already set
         }
-
-        _credentialsProvider.setCredentials(new AuthScope(httpRequest.getURI().getHost(), httpRequest.getURI().getPort()), creds);
-    }    
-
+    }
+    
+    /**
+     * Create an HTTP request with the appropriate method and body
+     */
+    private SimpleHttpRequest createRequest(String method, String url, String body, String contentType) {
+        // Determine the effective method
+        String effectiveMethod;
+        if (Strings.isBlank(method)) {
+            effectiveMethod = Strings.isEmpty(body) ? "GET" : "POST";
+        } else {
+            effectiveMethod = method;
+        }
+        
+        // Build the request
+        SimpleRequestBuilder builder = SimpleRequestBuilder.create(effectiveMethod).setUri(url);
+        
+        // Set body if provided
+        if (!Strings.isEmpty(body)) {
+            ContentType cType = ContentType.create(
+                    contentType != null ? contentType : "text/plain", "utf-8");
+            builder.setBody(body, cType);
+        }
+        
+        return builder.build();
+    }
+    
+    @Override
+    public CompletableFuture<String> makeSimpleRequestAsync(String urlStr, String method, Map<String, String> query,
+                              String username, String password,
+                              Map<String, String> headers, String contentType, String post,
+                              Integer connectTimeout, Integer readTimeout) {
+        
+        return makeRequestAsync(urlStr, method, query, username, password, headers, contentType, post, connectTimeout, readTimeout)
+                .thenApply(response -> {
+                    if (response.statusCode >= 200 && response.statusCode < 300) {
+                        return response.content;
+                    } else {
+                        throw new RuntimeException(String.format("Server returned '%s' with content %s", 
+                                response.statusCode + " " + response.reasonPhrase, 
+                                Strings.isEmpty(response.content) ? "<empty>" : org.nodel.json.JSONObject.quote(response.content)));
+                    }
+                });
+    }
+    
     @Override
     public void close() throws IOException {
-        Stream.safeClose(_httpClient);
-    }
-    
-    // static convenience methods
-    
-    /**
-     * Returns the local host name (does not throw exceptions).
-     */
-    private static String getLocalHostName() {
-        try {
-            return InetAddress.getLocalHost().getHostName();
-        } catch (UnknownHostException exc) {
-            throw new UnexpectedIOException(exc);
-        }
-    }
-    
-    // convenience instances
-    
-    /**
-     * Ignores all SSL issues
-     */
-    private static X509TrustManager IGNORE_SSL_TRUSTMANAGER = new X509TrustManager() {
-        
-        public void checkClientTrusted(X509Certificate[] chain, String authType) throws CertificateException { }
-        
-        public void checkServerTrusted(X509Certificate[] chain, String authType) throws CertificateException { }
-        
-        public X509Certificate[] getAcceptedIssuers() {
-            return new X509Certificate[0];
-        }
-        
-    };
-    
-    /**
-     * Is injected into stream to count bytes (vs characters)
-     */
-    private static class SafeCounter implements SharableMeasurementProvider {
-
-        private long _counter;
-
-        @Override
-        public long getMeasurement() {
-            return _counter;
-        }
-
-        @Override
-        public void set(long value) {
-            _counter = value;
-
-            if (_counter > MAX_ALLOWED)
-                throw new UnexpectedIOException("Too big - HTTP response over " + MAX_ALLOWED + " bytes is not allowed");
-        }
-
-        @Override
-        public void add(long value) {
-            _counter += value;
+        synchronized (_lock) {
+            // Close asynchronous client if initialized
+            if (_httpAsyncClient != null) {
+                _httpAsyncClient.close(CloseMode.GRACEFUL);
+                _httpAsyncClient = null;
+            }
             
-            if (_counter > MAX_ALLOWED)
-                throw new UnexpectedIOException("Too big - HTTP response over " + MAX_ALLOWED + " bytes is not allowed");
-        }
-
-        @Override
-        public void incr() {
-            _counter++;
+            // Shutdown executor service if initialized
+            if (_executor != null) {
+                _executor.shutdown();
+                try {
+                    // Wait for pending tasks to complete
+                    if (!_executor.awaitTermination(5, TimeUnit.SECONDS)) {
+                        _executor.shutdownNow();
+                    }
+                } catch (InterruptedException e) {
+                    _executor.shutdownNow();
+                    Thread.currentThread().interrupt();
+                }
+                _executor = null;
+            }
             
-            if (_counter > MAX_ALLOWED)
-                throw new UnexpectedIOException("Too big - HTTP response over " + MAX_ALLOWED + " bytes is not allowed");
+            _credentialsProvider = null;
+            _requestConfig = null;
         }
-
-        @Override
-        public void decr() {
-            _counter--;
-        }
-
     }
-
-    /**
-     *  A redirect strategy used to ignore all redirect directives i.e. will be manually handled
-     */
-    private static RedirectStrategy IGNORE_ALL_REDIRECTS = new RedirectStrategy() {
-
-        @Override
-        public boolean isRedirected(HttpRequest request, HttpResponse response, HttpContext context) throws ProtocolException {
-            return false;
-        }
-
-        @Override
-        public HttpUriRequest getRedirect(HttpRequest request, HttpResponse response, HttpContext context) throws ProtocolException {
-            return null;
-        }
-        
-    };
-    
 }
